Улучшить метод вывода поля interrupt регистра mcause; исправить недочёты rvmath связанные с длинной операндов

+++ hdr/rvmath.h
@@ -1,7 +1,7 @@
 #ifndef RVMATH
 #define RVMATH
 
-unsigned int i_mod(const unsigned int a, const unsigned int b);
-unsigned int i_div(const unsigned int a, const unsigned int b);
+unsigned long long i_mod(const unsigned long long a, const unsigned long long b);
+unsigned long long i_div(const unsigned long long a, const unsigned long long b);
 
 #endif
